Address review comments - Have separate view for "Send push" flow - Rename json files
