+++ src/v2/view-builder/views/SelectAuthenticatorVerifyView.js
@@ -5,8 +5,6 @@ export const Body = BaseForm.extend({
   title: function() {
     if (this.isPasswordRecoveryFlow())  {
       return loc('password.reset.title.generic', 'login');
-    } else if (this.isPushOnlyFlow()) {
-      return loc('oie.okta_verify.push.title', 'login');
     }
     return loc('oie.select.authenticators.verify.title', 'login');
   },
@@ -18,12 +16,6 @@ export const Body = BaseForm.extend({
   },
   isPasswordRecoveryFlow() {
     return this.options.appState.get('isPasswordRecovery');
-  },
-  isPushOnlyFlow() {
-    const uiSchema = this.options.currentViewState.uiSchema;
-    const methodType = Array.isArray(uiSchema) && uiSchema.find(schema => schema.name === 'authenticator.methodType');
-    const methodTypeOptions = methodType && methodType.options;
-    return Array.isArray(methodTypeOptions) && methodTypeOptions.length === 1 && methodTypeOptions[0].value === 'push';
   },
   noButtonBar: true,
 });
