+++ test/testcafe/framework/page-objects/ChallengeOktaVerifyPushPageObject.js
@@ -1,12 +1,12 @@
+import { Selector } from 'testcafe';
 import ChallengeFactorPageObject from './ChallengeFactorPageObject';
 
 const FORM_INFOBOX_WARNING = '.okta-form-infobox-warning';
 const FORM_INFOBOX_ERROR = '[data-se="o-form-error-container"] [data-se="callout"]';
 const RESEND_NUMBER_CHALLENGE_BUTTON = '.okta-form-infobox-warning .resend-number-challenge';
 const FORM_INFOBOX_ERROR_TITLE = '[data-se="o-form-error-container"] [data-se="callout"] > h3';
-const AUTO_CHALLENGE_CHECKBOX_SELECTOR = '[name="autoChallenge"]';
-const AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR = '[data-se-for-name="autoChallenge"]';
-
+const SUBTITLE_SELECTOR = '[data-se="o-form-explain"]';
+const FORM_SELECTOR = '.okta-verify-push-challenge';
 export default class ChallengeOktaVerifyPushPageObject extends ChallengeFactorPageObject {
   constructor(t) {
     super(t);
@@ -48,16 +48,29 @@ export default class ChallengeOktaVerifyPushPageObject extends ChallengeFactorPa
     return this.form.getElement(FORM_INFOBOX_WARNING);
   }
 
-  async autoChallengeInputExists() {
-    return this.form.elementExist(AUTO_CHALLENGE_CHECKBOX_SELECTOR);
+  subtitleExists() {
+    return this.form.elementExist(SUBTITLE_SELECTOR);
   }
 
-  getAutoChallengeCheckboxLabel() {
-    return this.form.getElement(AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR);
+  async autoChallengeInputExists(selector) {
+    return this.form.elementExist(selector);
   }
 
-  async clickAutoChallengeCheckbox() {
-    await this.t.click(this.form.getElement(AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR));
+  getAutoChallengeCheckboxLabel(selector) {
+    return this.form.getElement(selector);
+  }
+
+  async clickAutoChallengeCheckbox(selector) {
+    await this.t.click(this.form.getElement(selector));
+  }
+
+  async  formWithClassExist() {
+    const formCount = await Selector(FORM_SELECTOR).count;
+    return formCount === 1;
+  }
+
+  async clickSendPushButton() {
+    await this.t.click(this.getPushButton());
   }
 
 }
