+++ test/testcafe/spec/ChallengeOktaVerifyPush_spec.js
@@ -5,7 +5,10 @@ import { checkConsoleMessages } from '../framework/shared';
 
 import pushPoll from '../../../playground/mocks/data/idp/idx/authenticator-verification-okta-verify-push';
 import success from '../../../playground/mocks/data/idp/idx/success';
-import sendPushPoll from '../../../playground/mocks/data/idp/idx/challenge-with-push-notification';
+import sendPushPoll from '../../../playground/mocks/data/idp/idx/authenticator-verification-okta-verify-push-autoChallenge';
+
+const AUTO_CHALLENGE_CHECKBOX_SELECTOR = '[name="autoChallenge"]';
+const AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR = '[data-se-for-name="autoChallenge"]';
 
 const logger = RequestLogger(/challenge|challenge\/poll|authenticators\/poll/,
   {
@@ -61,7 +64,7 @@ test
     await t.expect(a11ySpan.textContent).contains('Push notification sent');
     await t.expect(pushBtn.hasClass('link-button-disabled')).ok();
     await t.expect(pageTitle).contains('Get a push notification');
-    await t.expect(await challengeOktaVerifyPushPageObject.autoChallengeInputExists()).notOk();
+    await t.expect(await challengeOktaVerifyPushPageObject.autoChallengeInputExists(AUTO_CHALLENGE_CHECKBOX_SELECTOR)).notOk();
 
     // Verify links
     await t.expect(await challengeOktaVerifyPushPageObject.switchAuthenticatorLinkExists()).ok();
@@ -83,17 +86,17 @@ test
     const pageTitle = challengeOktaVerifyPushPageObject.getFormTitle();
     const pushBtn = challengeOktaVerifyPushPageObject.getPushButton();
     const a11ySpan = challengeOktaVerifyPushPageObject.getA11ySpan();
-    const checkboxLabel = challengeOktaVerifyPushPageObject.getAutoChallengeCheckboxLabel();
+    const checkboxLabel = challengeOktaVerifyPushPageObject.getAutoChallengeCheckboxLabel(AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR);
     await t.expect(pushBtn.textContent).contains('Push notification sent');
     await t.expect(a11ySpan.textContent).contains('Push notification sent');
     await t.expect(pushBtn.hasClass('link-button-disabled')).ok();
     await t.expect(pageTitle).contains('Get a push notification');
-    await t.expect(await challengeOktaVerifyPushPageObject.autoChallengeInputExists()).ok();
+    await t.expect(await challengeOktaVerifyPushPageObject.autoChallengeInputExists(AUTO_CHALLENGE_CHECKBOX_SELECTOR)).ok();
     await t.expect(checkboxLabel.hasClass('checked')).ok();
     await t.expect(checkboxLabel.textContent).eql('Send push automatically');
 
     // unselect checkbox on click
-    await challengeOktaVerifyPushPageObject.clickAutoChallengeCheckbox();
+    await challengeOktaVerifyPushPageObject.clickAutoChallengeCheckbox(AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR);
     await t.expect(checkboxLabel.hasClass('checked')).notOk();
 
     // Verify links
@@ -106,8 +109,8 @@ test
 test
   .requestHooks(logger, sendPushWaitMock)('should call polling API and checkbox should be clickable after polling started', async t => {
     const challengeOktaVerifyPushPageObject = await setup(t);
-    const checkboxLabel = challengeOktaVerifyPushPageObject.getAutoChallengeCheckboxLabel();
-    await t.expect(await challengeOktaVerifyPushPageObject.autoChallengeInputExists()).ok();
+    const checkboxLabel = challengeOktaVerifyPushPageObject.getAutoChallengeCheckboxLabel(AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR);
+    await t.expect(await challengeOktaVerifyPushPageObject.autoChallengeInputExists(AUTO_CHALLENGE_CHECKBOX_SELECTOR)).ok();
     await t.expect(checkboxLabel.hasClass('checked')).ok();
     await t.expect(checkboxLabel.textContent).eql('Send push automatically');
     // wait for polling to start
@@ -129,7 +132,7 @@ test
     await t.expect(answerRequestUrl).eql('http://localhost:3000/idp/idx/authenticators/poll');
 
     // unselect checkbox on click
-    await challengeOktaVerifyPushPageObject.clickAutoChallengeCheckbox();
+    await challengeOktaVerifyPushPageObject.clickAutoChallengeCheckbox(AUTO_CHALLENGE_CHECKBOX_LABEL_SELECTOR);
     await t.expect(checkboxLabel.hasClass('checked')).notOk();
   });
